app/src/main/java/com/mgaetan89/showsrage/activity/HistoryActivity.java
package com.mgaetan89.showsrage.activity;

import android.os.Bundle;

import com.mgaetan89.showsrage.R;
import com.mgaetan89.showsrage.fragment.HistoryFragment;

public class HistoryActivity extends BaseActivity {
	@Override
	protected int getSelectedMenuId() {
		return R.id.menu_history;
	}

	@Override
	protected int getTitleResourceId() {
		return R.string.history;
	}

	@Override
	protected void onCreate(Bundle savedInstanceState) {
		super.onCreate(savedInstanceState);

		if (savedInstanceState == null) {
			this.getSupportFragmentManager().beginTransaction()
					.replace(R.id.content, new HistoryFragment())
					.commit();
		}
	}

	@Override
	protected void onResume() {
		super.onResume();

		this.displayHomeAsUp(false);
	}
}
