app/src/test/java/com/mgaetan89/showsrage/activity/HistoryActivityTest.java
package com.mgaetan89.showsrage.activity;

import com.mgaetan89.showsrage.R;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;

import static org.assertj.core.api.Assertions.assertThat;

public class HistoryActivityTest {
	private HistoryActivity activity;

	@Before
	public void before() {
		this.activity = new HistoryActivity();
	}

	@Test
	public void getSelectedMenuId() {
		assertThat(this.activity.getSelectedMenuId()).isEqualTo(R.id.menu_history);
	}

	@Test
	public void getTitleResourceId() {
		assertThat(this.activity.getTitleResourceId()).isEqualTo(R.string.history);
	}

	@After
	public void after() {
		this.activity = null;
	}
}